requirements.txt
# =========================================================
# Core Airflow dependencies (não incluir apache-airflow aqui)
# =========================================================

# =========================================================
# Data & Processing
# =========================================================
pyspark==3.5.1
pandas==2.2.2
pyarrow==15.0.2
fastparquet==2024.5.0

# =========================================================
# Observability & Logging
# =========================================================
mlflow==2.14.3
prometheus-client==0.20.0
requests==2.32.3
python-json-logger==2.0.7

# =========================================================
# Configuration & Utilities
# =========================================================
PyYAML==6.0.2
tenacity==8.2.2  # downgrade leve e compatível
attrs==23.2.0    # downgrade leve e compatível

# =========================================================
# Testing & QA
# =========================================================
pytest==8.3.2
pytest-mock==3.14.0
pytest-cov==5.0.0
